server/agent/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth
boto3
pydantic-settings

pydantic_ai